Fixed problem with JDO test suite (introduced with commit set 6424) (CASTOR-1706).

+++ castor/trunk/src/main/java/org/exolab/castor/jdo/engine/JDOClassDescriptor.java
@@ -124,7 +124,7 @@ public class JDOClassDescriptor extends ClassDescriptorImpl {
 
         CacheTypeMapping cacheMapping = getMapping().getCacheTypeMapping();
         if (cacheMapping != null) {
-            String capacity = Integer.toString(cacheMapping.getCapacity());
+            String capacity = Long.toString(cacheMapping.getCapacity());
             _cacheParams.put(CountLimited.PARAM_CAPACITY, capacity);
             _cacheParams.put(TimeLimited.PARAM_TTL, capacity);
